refactored PeakData into two classes and updated their usages with unit tests
